% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nba_team_game_models.R
\name{build_nba_team_game_models}
\alias{build_nba_team_game_models}
\title{Build NBA team-game models}
\usage{
build_nba_team_game_models(modeling_data = NULL, output_path = NULL,
  n_cores = round(parallel::detectCores()/2))
}
\arguments{
\item{modeling_data}{output of \code{preprocess_nba_team_games()}}

\item{output_path}{character path where models should be saved}

\item{n_cores}{number of cores to use for parallelization}
}
\value{
a list of models, an ensemble, and data
}
\description{
Build NBA team-game models
}
