% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_schema_nba.R
\name{create_db_schema_nba}
\alias{create_db_schema_nba}
\title{Create nba schema in database}
\usage{
create_db_schema_nba(db_connection = NULL, create_schema_sql = NULL)
}
\arguments{
\item{db_connection}{connection object created with \code{DBI::dbConnect()}}

\item{create_schema_sql}{character path to sql file for creating nba schema}
}
\value{
nothing
}
\description{
Create nba schema in database
}
