% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table_nba_boxscores.R
\name{create_db_table_nba_boxscores}
\alias{create_db_table_nba_boxscores}
\title{Create nba.boxscores table in database}
\usage{
create_db_table_nba_boxscores(seasons = NULL, db_connection = NULL,
  create_table_sql = NULL)
}
\arguments{
\item{seasons}{seasons to add for initial import}

\item{db_connection}{connection object created with \code{DBI::dbConnect()}}

\item{create_table_sql}{character path to sql file for creating nba.boxscores table}
}
\value{
row count
}
\description{
Create nba.boxscores table in database
}
