% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_nba_team_games.R
\name{enrich_nba_team_games}
\alias{enrich_nba_team_games}
\title{Enrich the bejesus out of nba team-game data}
\usage{
enrich_nba_team_games(team_games = NULL)
}
\arguments{
\item{team_games}{data.frame of nba team game data}
}
\value{
a much, much wider data.frame of nba team game data
}
\description{
Enrich the bejesus out of nba team-game data
}
