% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_nba_team_games_for_scoring.R
\name{enrich_nba_team_games_for_scoring}
\alias{enrich_nba_team_games_for_scoring}
\title{Enrich NBA team-game data for scoring}
\usage{
enrich_nba_team_games_for_scoring(game_slate = NULL,
  team_games_for_scoring = NULL)
}
\arguments{
\item{game_slate}{data.frame output of \code{nbastatstools::get_nba_game_slate()}}

\item{team_games_for_scoring}{data.frame of team-games-to-be-scored}
}
\value{
data.frame
}
\description{
Enrich NBA team-game data for scoring
}
