% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_game_slate.R
\name{get_nba_game_slate}
\alias{get_nba_game_slate}
\title{Get slate of nba games for a date}
\usage{
get_nba_game_slate(game_date = Sys.Date())
}
\arguments{
\item{game_date}{date or character formatted yyyy-mm-dd}
}
\value{
slate of games for a date
}
\description{
Get slate of nba games for a date
}
