% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_playbyplay.R
\name{get_nba_playbyplay}
\alias{get_nba_playbyplay}
\title{get play-by-play for games from nba stats api}
\usage{
get_nba_playbyplay(gameids = NULL, sleep = TRUE)
}
\arguments{
\item{gameid}{character string usually represented in data as \code{GAME_ID}}
}
\value{
a list of data.frames
}
\description{
get play-by-play for games from nba stats api
}
