% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_player_profiles.R
\name{get_nba_player_profiles}
\alias{get_nba_player_profiles}
\title{get player profiles from nba stats api}
\usage{
get_nba_player_profiles(playerids = NULL, sleep = TRUE)
}
\arguments{
\item{playerids}{character vector of playerids}

\item{sleep}{logical should there be \code{Sys.sleep()} between playerids; default is \code{TRUE}}
}
\value{
data.frame of player profiles
}
\description{
get player profiles from nba stats api
}
