% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nba_season_team_rosters.R
\name{get_nba_season_team_rosters}
\alias{get_nba_season_team_rosters}
\title{Query the commonteamroster endpoint from nba stats}
\usage{
get_nba_season_team_rosters(seasons = NULL, teamids = NULL,
  sleep = TRUE)
}
\arguments{
\item{seasons}{numeric seasons to grab (e.g. 2000)}

\item{teamids}{character that looks like a numeric but could have leading zeroes, teamids to grab, usually the TEAM_ID field in a boxscore}

\item{sleep}{logical sleep between API calls}
}
\value{
list of team rosters; each entry will have players and coaches if available
}
\description{
Query the commonteamroster endpoint from nba stats
}
