% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_nba_boxscores.R
\name{get_new_nba_boxscores}
\alias{get_new_nba_boxscores}
\title{Get new boxscores from nba stats api}
\usage{
get_new_nba_boxscores(db_connection = NULL, boxscores_table = NULL,
  output_dir = NULL)
}
\arguments{
\item{db_connection}{RPostgreSQL connection object}

\item{boxscores_table}{character location of boxscores table in db}

\item{output_dir}{character path to output new boxscores}
}
\value{
save a .csv to output_dir
}
\description{
Get new boxscores from nba stats api
}
