% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_team_season_cumulative_leaders.R
\name{identify_team_season_cumulative_leaders}
\alias{identify_team_season_cumulative_leaders}
\title{Identify players who are leading team in stat categories for a season coming into a game}
\usage{
identify_team_season_cumulative_leaders(boxscores = NULL,
  stat_list = c("min", "pts", "fgthreem", "reb", "ast", "plus_minus"))
}
\arguments{
\item{boxscores}{data.frame of boxscore info}

\item{stat_list}{list of string names of fields to calc leaders on}
}
\value{
data.frame of team-season-cumulative-leaders
}
\description{
Identify players who are leading team in stat categories for a season coming into a game
}
