% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_team_season_leaders_in_lineup.R
\name{identify_team_season_leaders_in_lineup}
\alias{identify_team_season_leaders_in_lineup}
\title{Indicate if a team-season-leader in a stat category is in lineup for game}
\usage{
identify_team_season_leaders_in_lineup(boxscores = NULL)
}
\arguments{
\item{boxscores}{data.frame of boxscores}
}
\value{
data.frame
}
\description{
Indicate if a team-season-leader in a stat category is in lineup for game
}
