% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nba_team_games.R
\name{make_nba_team_games}
\alias{make_nba_team_games}
\title{make a team-games table from nba stats api boxscores}
\usage{
make_nba_team_games(boxscores = NULL)
}
\arguments{
\item{boxscores}{data.frame of boxscores from nba stats api}
}
\value{
data.frame of team_game info for easy joining
}
\description{
make a team-games table from nba stats api boxscores
}
