% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_team_game_player_rosters.R
\name{make_team_game_player_rosters}
\alias{make_team_game_player_rosters}
\title{List active players by games and teams}
\usage{
make_team_game_player_rosters(boxscores = NULL)
}
\arguments{
\item{boxscores}{data.frame of boxscores}
}
\value{
list of games with lists of rosters
}
\description{
List active players by games and teams
}
