% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_nba_team_games.R
\name{preprocess_nba_team_games}
\alias{preprocess_nba_team_games}
\title{Preprocess team-game data for model-building and scoring}
\usage{
preprocess_nba_team_games(enriched_team_games = NULL, models = NULL,
  outcomes = NULL, objects_output_path = NULL)
}
\arguments{
\item{enriched_team_games}{data.frame with all features engineered}

\item{models}{models to build}

\item{outcomes}{outcomes to predict for the models to build}
}
\value{
list of pre-processing objects and data for building models
}
\description{
Preprocess team-game data for model-building and scoring
}
