% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_nba_team_games_for_scoring.R
\name{preprocess_nba_team_games_for_scoring}
\alias{preprocess_nba_team_games_for_scoring}
\title{Apply preprocessing and dummy-variable coding in prep for scoring}
\usage{

  preprocess_nba_team_games_for_scoring(enriched_team_games_for_scoring = NULL,
  modeling_objects_path = NULL)
}
\arguments{
\item{enriched_team_games_for_scoring}{data.frame of of nba team-game data}

\item{modeling_objects_path}{path to object that contains modeling objects}
}
\value{
list of data for modeling
}
\description{
Apply preprocessing and dummy-variable coding in prep for scoring
}
