% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_nba_team_games.R
\name{score_nba_team_games}
\alias{score_nba_team_games}
\title{Generate NBA team-game scores using selected models}
\usage{
score_nba_team_games(preprocessed_team_games_for_scoring = NULL,
  model_path = NULL)
}
\arguments{
\item{preprocessed_team_games_for_scoring}{list of data.frames as model inputs}

\item{model_path}{character string path to model objects}
}
\value{
list of data.frames with model outputs
}
\description{
Generate NBA team-game scores using selected models
}
