% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_nba_boxscores.R
\name{tidy_nba_boxscores}
\alias{tidy_nba_boxscores}
\title{tidy boxscores from nba stats api}
\usage{
tidy_nba_boxscores(boxscores = NULL)
}
\arguments{
\item{boxscores}{data.frame with boxscore data from nba stats api}
}
\value{
tidy data.frame
}
\description{
tidy boxscores from nba stats api
}
