% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_nba_game_slate.R
\name{tidy_nba_game_slate}
\alias{tidy_nba_game_slate}
\title{Tidy game_slate}
\usage{
tidy_nba_game_slate(game_slate = NULL)
}
\arguments{
\item{game_slate}{data.frame output of \code{nbastatstools::get_nba_game_slate()}}
}
\value{
data.frame
}
\description{
Tidy game_slate
}
