% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_nba_playbyplay.R
\name{tidy_nba_playbyplay}
\alias{tidy_nba_playbyplay}
\title{tidy playbyplay from nba stats api}
\usage{
tidy_nba_playbyplay(playbyplay = NULL)
}
\arguments{
\item{playbyplay}{data.frame with playbyplay data from nba stats api}
}
\value{
tidy data.frame
}
\description{
tidy playbyplay from nba stats api
}
